requirements.txt
influxdb
pyserial
